// src/interfaces/score.interface.ts
export interface IScore {
  ID: string;
  FullName: string;
  Account: string;
  ASM: number;
  HTML: number;
  CSS: number;
  Quiz3: number;
  Quiz4: number;
  Quiz5: number;
  Quiz6: number;
  AvgQuiz: number;
  Practice1: number;
  Practice2: number;
  Practice3: number;
  AvgASM: number;
  QuizFinal: number;
  Audit: number;
  PracticeFinal: number;
  FinalModule: number;
  GPAModule: number;
  LevelModule: number;
  Mock: number;
  MockFinalModule: number;
  MockGPAModule: number;
  MockLevelModule: number;
  Status: boolean;
  MockStatus: boolean;
}